+++ search-service/src/main/java/com/example/searchservice/commission/service/CommissionService.java
@@ -1,10 +1,21 @@
 package com.example.searchservice.commission.service;
 
 import com.example.searchservice.commission.dto.CommissionResponseDto;
+import com.example.searchservice.commission.entity.CommissionDocumentEntity;
 import com.example.searchservice.common.vo.SearchScope;
+import java.util.List;
 import org.springframework.data.domain.Page;
 
 public interface CommissionService {
 
     public Page<CommissionResponseDto> search(String query, SearchScope scope, int page, int size);
+
+    public void saveAll(List<CommissionDocumentEntity> commissions);
+
+    public void save(CommissionDocumentEntity commission);
+
+    public void update(CommissionDocumentEntity commission);
+
+    public void delete(String code);
+
 }
