+++ search-service/src/main/java/com/example/searchservice/tag/service/TagService.java
@@ -1,9 +1,12 @@
 package com.example.searchservice.tag.service;
 
 import com.example.searchservice.tag.dto.TagResponseDto;
+import com.example.searchservice.tag.service.dto.ProfileTagDto;
 import java.util.List;
 
 public interface TagService {
     public List<TagResponseDto> getSuggestions(String prefix, int size);
+
+    public void saveAll(List<ProfileTagDto> tags);
 }
